re-factor - add same menu for tpe def grid and entity views refinements to publish ui.
